services/lesson-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.3
python-jose[cryptography]==3.3.0
pydantic==2.10.0
python-multipart==0.0.12
aiofiles==24.1.0
